additional fixes, example output files

+++ src/AllSplits/AllSplits.hpp
@@ -24,12 +24,13 @@ class AllSplits : public Algorithm{
     matrix_weight_t m_ext_cost_tab;
     
     //Map binding the binary key of a state to a cost
-    std::unordered_map<unsigned long long, cost_t> m_cost_map;
+    //std::unordered_map<unsigned long long, cost_t> m_cost_map;
+    std::unordered_map<double, cost_t> m_cost_map;
 
     //unordered_map<unsigned long long, int> P1; //la liste des ordres S1
-    std::unordered_map<unsigned long long, int> m_order_map_1;
+    std::unordered_map<double, int> m_order_map_1;
 
-    std::unordered_map<unsigned long long, int> m_order_map_2; //la liste des ordres S2
+    std::unordered_map<double, int> m_order_map_2; //la liste des ordres S2
 
     //G : un tableau de taille n+1*n, la matrice d'adjacence + la colonne A de départ. M[i][j] = G[i*size+j]
     //S : Une liste de sommet, les sommets sélectionné pour cet état
@@ -53,9 +54,9 @@ class AllSplits : public Algorithm{
     cost_t produit_sortant(vector_vertexID_t const& state, matrix_weight_t const& ext_cost_list);
 
     //converti un ensemble de sommets en un entier pouvant être stocké dans une map
-    unsigned long long convert(vector_vertexID_t const& state);
+    double convert(vector_vertexID_t const& state);
     //converti une clé en l'ensemble de sommets correspondant
-    vector_vertexID_t recover(unsigned long long key);
+    vector_vertexID_t recover(double key);
 
     void display_order(vector_vertexID_t const& state);
     void display_order();
